backend/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.3.0

# Rate limiting
slowapi>=0.1.9

# File uploads
python-multipart>=0.0.6

# Optional - install separately if needed:
# pandas>=2.1.0
# geopandas>=0.14.0
# reportlab>=4.0.0
# anthropic>=0.18.0
